state-res: Enforce integer PLs for room v10 on custom types
According to MSC3667
